hacknu24mitnick/backend/requirements.txt
Django==5.0.4
python-slugify==8.0.4